requirements.txt
speech_recognition
PyAudio
pyttsx3
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
nltk
customtkinter
pillow
python-dotenv
email-validator
requests
pickle-mixin
python-dotenv>=0.19.0
nltk>=3.6.0
sounddevice>=0.4.1
selenium>=4.0.0
webdriver-manager>=3.8.0
